Simplistic2ndpop
Simplistic method to adding the second population.

Define a second set of input parameters into the theta, along with a ratio of the two populations.

Test_fitting creates a template using an input theta dict, and then attempts to reset theta and find the best fit solution. This should in theory test if the inputs are recreated. However, there are many degeneracies in getting out the exact inputs.